src/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.1.0
fastapi==0.115.2
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
gradio==4.44.1
gradio-client==1.3.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.8
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
jinja2==3.1.4
jupyter-client==8.6.2
jupyter-core==5.7.2
kiwisolver==1.4.7
llvmlite==0.43.0
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
orjson==3.10.9
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydub==0.25.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-multipart==0.0.12
pytz==2024.2
pyyaml==6.0.2
pyzmq==26.2.0
regex==2024.9.11
requests==2.32.3
rich==13.9.2
ruff==0.7.0
semantic-version==2.10.0
setuptools==74.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.1
tiktoken==0.8.0
tomlkit==0.12.0
torch==2.5.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.32.0
wcwidth==0.2.13
websockets==12.0
